eSams/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
colorama==0.4.6
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dlib @ file:///C:/Users/MWINYOGLEE/Downloads/dlib-for-python3.10/dlib-for-python3.10/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
html5lib==1.1
idna==3.4
keras==2.13.1
keras-facenet==0.3.2
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
mtcnn==0.1.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyphen==0.14.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.15.0
